Support built-in Sodium extension in PHP 7.2+ on Ubuntu (#5)
* Fix unbound variable error on Mac OS X

* Support built-in Sodium extension in PHP 7.2+ on Ubuntu

Problem:
Attempting to use this package on Ubuntu with PHP 7.2+ results in Composer complaining that the installed Sodium extension does not match the ^2.0 requirements.

Cause:
After PHP 7.1, the Sodium extension is built in to PHP. On Ubuntu, the version string for the package is a crazy ubuntu string instead of a typical semver string. As such, composer cannot determine that the version installed is actually 2.0 or greater.

Fix:
Although not ideal, I modified the composer.json to require any version of the extension. We do still literally need 2.0, but since Ubuntu formats the extension version strangely, there's no way to enforce it without breaking compatability with Ubuntu.

* Enabled Docker layer caching on CircleCI

* Fix duplicate configure_git steps in CircleCI config

* Combine pull-merge with restore in CircleCI

* Run tests against all minor versions of PHP 7